DM:tools: add APIs for debugger
This patch is the initialization patch for some APIs for debugger.

Usercrash works as C/S model. Debugger runs as the client of usercrash.
File client.c provides the APIs for the connection ops between debugger
and usercrashd. File crash_dump.c provides the APIs for debugger to
dump some crash information.

Signed-off-by: LIU Xiaojing <xiaojing.liu@intel.com>
Signed-off-by: CHEN Gang <gang.c.chen@intel.com>